--- src/index.test.ts
@@ -1,7 +1,0 @@
-import { describe, it, expect } from 'vitest';
-
-describe('Focus Trap Class', () => {
-	it('', () => {
-		expect(1 + 2).toBe(3);
-	});
-});
