+++ tests/test.ts
@@ -4,3 +4,7 @@ test('index page has expected h1', async ({ page }) => {
 	await page.goto('/');
 	await expect(page.getByRole('heading', { name: 'Welcome to SvelteKit' })).toBeVisible();
 });
+
+test('use:focusTrap', async ({ page }) => {
+	await page.goto('/');
+});
